+++ src/Jadob/Dashboard/Action/DashboardAction.php
@@ -104,8 +104,35 @@ class DashboardAction
         if ($action === ActionType::OPERATION) {
             return $this->handleOperation($request, $context);
         }
+
+        if($action === ActionType::BATCH_OPERATION) {
+            return $this->handleBatchOperation($request, $context);
+        }
     }
 
+
+    protected function handleBatchOperation(Request $request, DashboardContextInterface $context): Response
+    {
+        $this->logger->debug('handleOperation invoked');
+        $objectFqcn = $request->query->get(QueryStringParamName::OBJECT_NAME);
+        $operationName = $request->request->get('operation');
+        $managedObjectConfiguration = $this->configuration->getManagedObjectConfiguration($objectFqcn);
+
+        $this->logger->debug('Getting information about operation');
+        $operation = $managedObjectConfiguration->getListConfiguration()->getOperation($operationName);
+        $this->logger->debug('Getting object from persistence');
+        foreach ($request->request->get('id') as $objectId) {
+            $object = $this->doctrineOrmObjectManager->getOneById($objectFqcn, $objectId);
+            $this->logger->debug(sprintf('Continuing to invoke an operation "%s"', $operationName));
+            $this->operationHandler->processOperation($operation, $object, $context);
+            $this->logger->debug(sprintf('Operation "%s" invoked, returning to list view.', $operationName));
+        }
+
+        if($request->server->has('HTTP_REFERER')) {
+            return new RedirectResponse($request->server->get('HTTP_REFERER'));
+        }
+        return new RedirectResponse($this->pathGenerator->getPathForObjectList($objectFqcn));
+    }
     /**
      * @param Request $request
      * @return Response
@@ -186,6 +213,7 @@ class DashboardAction
                 $list = $objects;
             }
 
+
             return new Response(
                 $this->twig->render(
                     '@JadobDashboard/crud/list.html.twig', [
@@ -245,9 +273,11 @@ class DashboardAction
                 throw new DashboardException('There is no way to create a form.');
             }
 
+            if ($form === null) {
+                throw new RuntimeException('Form factory does not returned a Form!');
+            }
 
             $form->handleRequest($request);
-
             if ($form->isSubmitted() && $form->isValid()) {
                 /** @var object $createdObject */
                 $createdObject = $form->getData();