+++ src/Jadob/Dashboard/ActionType.php
@@ -8,6 +8,7 @@ class ActionType
 {
     public const IMPORT = 'import';
     public const OPERATION = 'operation';
+    public const BATCH_OPERATION = 'batch_operation';
     public const CRUD = 'crud';
     public const DASHBOARD = 'dashboard';