+++ src/Jadob/Security/Auth/UserProviderInterface.php
@@ -1,27 +1,19 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Jadob\Security\Auth;
 
 use Jadob\Security\Auth\User\UserInterface;
 
 /**
- * Interface UserProviderInterface
- *
- * @package Jadob\Security\Auth
- * @author  pizzaminded <mikolajczajkowsky@gmail.com>
+ * @author pizzaminded <mikolajczajkowsky@gmail.com>
+ * @license MIT
  */
 interface UserProviderInterface
 {
 
-    /**
-     * @param  $name
-     * @return null|UserInterface
-     */
     public function getUserByUsername(string $name): ?UserInterface;
 
-    /**
-     * @param  $id
-     * @return null|UserInterface
-     */
-    public function getUserById($id): ?UserInterface;
+    public function getUserById(string $id): ?UserInterface;
 }