+++ src/Jadob/Security/Supervisor/EventListener/AuthenticatorListener.php
@@ -72,7 +72,7 @@ class AuthenticatorListener implements ListenerProviderInterface, ListenerProvid
 
         //At first, handle stateless
         if ($requestSupervisor->isStateless()) {
-            $response = $this->handleStatelessRequest($event->getRequest(), $requestSupervisor);
+            $response = $this->handleStatelessRequest($event->getContext(), $requestSupervisor);
 
             if ($response !== null) {
                 $event->setResponse($response);
@@ -93,12 +93,13 @@ class AuthenticatorListener implements ListenerProviderInterface, ListenerProvid
 
     /**
      * @TODO: this must be moved to supervisor
-     * @param Request $request
+     * @param RequestContext $request
      * @param RequestSupervisorInterface $supervisor
      * @return Response|null
      */
-    protected function handleStatelessRequest(Request $request, RequestSupervisorInterface $supervisor): ?Response
+    protected function handleStatelessRequest(RequestContext $context, RequestSupervisorInterface $supervisor): ?Response
     {
+        $request = $context->getRequest();
         try {
             $credentials = $supervisor->extractCredentialsFromRequest($request);
 
@@ -117,9 +118,11 @@ class AuthenticatorListener implements ListenerProviderInterface, ListenerProvid
                 throw InvalidCredentialsException::invalidCredentials();
             }
         } catch (AuthenticationException $exception) {
+            $request->attributes->set(RequestAttribute::AUTHENTICATION_FAIL_REASON, $exception->getMessage());
             return $supervisor->handleAuthenticationFailure($exception, $request);
         }
 
+        $context->setUser($user);
         $this->identityStorage->setUser($user, $request->getSession(), get_class($supervisor));
         return $supervisor->handleAuthenticationSuccess($request, $user);
     }
@@ -161,6 +164,7 @@ class AuthenticatorListener implements ListenerProviderInterface, ListenerProvid
                 }
 
             } catch (AuthenticationException $exception) {
+                $request->attributes->set(RequestAttribute::AUTHENTICATION_FAIL_REASON, $exception->getMessage());
                 return $supervisor->handleAuthenticationFailure($exception, $request);
             }